client.app.store: Replace received None serial with ZERO_TID.
ZODB is inconsistent about object creation:
- database root is created with serial=None
- any other object is created with serial=ZERO_TID (aka z64 in ZODB code)
The meaning is the same on storage nodes (ignored when no existing
revision, will conflict otherwise), but the meaning is slightly different
for client cache.
In tpc_finish, objects at stored serial will be invalidated. For objects
stored with ZERO_TID, cache will take the lookup as a "load at this TID".
But for objects stored with None serial, cache would take the lookup as a
"load latest revision", which - although it should not hit as these are
new objects anyway - does not correspond to the intention.