lib.protocol: Include base TID in invalidation notifications.
For now, it only allows checking that invalidated cache entry is the
intended one.
It is intended to then be used to be able to process invalidations while
there are in-flight load calls, and accurately update their next_tid in
case of race-conditions, in turn allowing removal of load lock.