lib.connection: Use a more efficient data structure for request queue.
Packets being mostly processed in chronological timeout order, using an
ordered data structure is more efficient than a mapping by removing the
need for next_timeout housekeeping on most responses.
As typical operations are:
- peeking at both ends + (ap|pre)pending (request with normal or short
  timeout)
- popping first item (response received in chronological timeout order)
- peeking at first item (to know next timeout)
a collections.deque is more efficient than a list.

XXX: is next timeout always on the first pending queue ?
XXX: on timeout, is it always the first item of the first pending queue
which timed out ?